requirements-project.txt
aenum==3.1.15
agate==1.9.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
airbyte-api==0.52.1
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
apache-airflow==2.10.2
apache-airflow-providers-airbyte==4.0.0
apache-airflow-providers-celery==3.8.1
apache-airflow-providers-common-compat==1.2.1
apache-airflow-providers-common-io==1.4.0
apache-airflow-providers-common-sql==1.19.0
apache-airflow-providers-fab==1.3.0
apache-airflow-providers-ftp==3.11.0
apache-airflow-providers-google==10.24.0
apache-airflow-providers-http==4.13.0
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-mongo==4.2.2
apache-airflow-providers-postgres==5.13.1
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.9.0
apispec==6.6.1
argcomplete==3.5.0
asgiref==3.8.1
astro-sdk-python==1.8.1
astronomer-cosmos==1.6.0
astronomer-providers==1.19.4
attrs==24.2.0
azure-core==1.32.0
azure-storage-blob==12.19.0
babel==2.16.0
backoff==2.2.1
billiard==4.2.0
blinker==1.8.2
cachelib==0.9.0
cachetools==5.5.0
cattrs==24.1.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==6.8.2
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==3.0.3
cryptography==42.0.8
daff==1.3.46
dataclasses-json==0.6.7
db-dtypes==1.3.0
dbt-adapters==1.8.0
dbt-bigquery==1.8.2
dbt-common==1.11.0
dbt-core==1.8.7
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.5.1
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
dill==0.3.1.1
distlib==0.3.9
dnspython==2.6.1
docstring_parser==0.16
docutils==0.16
email_validator==2.2.0
Faker==20.1.0
filelock==3.16.1
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flower==2.0.1
frozenlist==1.4.1
fsspec==2024.10.0
gcloud-aio-auth==5.3.2
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.3.0
gcsfs==2024.10.0
google-ads==25.1.0
google-analytics-admin==0.23.2
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.71.1
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-automl==2.14.1
google-cloud-batch==0.17.31
google-cloud-bigquery==3.26.0
google-cloud-bigquery-datatransfer==3.17.1
google-cloud-bigtable==2.26.0
google-cloud-build==3.27.0
google-cloud-compute==1.20.1
google-cloud-container==2.53.0
google-cloud-core==2.4.1
google-cloud-datacatalog==3.21.1
google-cloud-dataflow-client==0.8.13
google-cloud-dataform==0.5.13
google-cloud-dataplex==2.3.1
google-cloud-dataproc==5.15.1
google-cloud-dataproc-metastore==1.16.0
google-cloud-dlp==3.25.0
google-cloud-kms==3.1.0
google-cloud-language==2.15.0
google-cloud-logging==3.11.3
google-cloud-memcache==1.10.0
google-cloud-monitoring==2.23.0
google-cloud-orchestration-airflow==1.15.0
google-cloud-os-login==2.15.0
google-cloud-pubsub==2.26.1
google-cloud-redis==2.16.0
google-cloud-resource-manager==1.13.0
google-cloud-run==0.10.10
google-cloud-secret-manager==2.21.0
google-cloud-spanner==3.49.1
google-cloud-speech==2.28.0
google-cloud-storage==2.18.2
google-cloud-storage-transfer==1.13.0
google-cloud-tasks==2.17.0
google-cloud-texttospeech==2.21.0
google-cloud-translate==3.17.0
google-cloud-videointelligence==2.14.0
google-cloud-vision==3.8.0
google-cloud-workflows==1.15.0
google-crc32c==1.6.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.0
grpc-google-iam-v1==0.13.1
grpc-interceptor==0.15.4
grpcio==1.66.1
grpcio-gcp==0.2.2
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
humanize==4.10.0
idna==3.10
immutabledict==4.2.0
importlib-metadata==6.11.0
importlib_resources==6.4.5
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
json-merge-patch==0.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.1
lazy-object-proxy==1.10.0
leather==0.4.0
limits==3.13.0
linkify-it-py==2.0.3
load-dotenv==0.1.0
lockfile==0.12.2
Logbook==1.5.3
looker-sdk==24.18.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mashumaro==3.14
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
minimal-snowplow-tracker==0.0.2
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
packaging==24.1
pandas==2.1.4
pandas-gbq==0.24.0
parsedatetime==2.6
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.3.6
pluggy==1.5.0
prison==0.2.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==4.25.4
psutil==6.0.0
psycopg2-binary==2.9.9
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.0
pycparser==2.22
pycpfcnpj==1.7.0
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydata-google-auth==1.8.2
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.10.1
pyOpenSSL==24.2.1
pyparsing==3.2.0
pytest==7.4.3
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-frontmatter==1.1.0
python-nvd3==0.16.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.8.1
rich-argparse==1.5.2
rpds-py==0.20.0
rsa==4.9
setproctitle==1.3.3
setuptools==75.3.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
SQLAlchemy==1.4.54
sqlalchemy-bigquery==1.12.0
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.7.0
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.15.0
tornado==6.4.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.5
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
virtualenv==20.27.1
wcwidth==0.2.13
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.11.1
zipp==3.20.2
